fix: corregir importaciones que no se usar

+++ src/main/java/com/Proyecto/La_Tertulia/controller/SalesController.java
@@ -115,7 +115,7 @@ public class SalesController {
         return "SaleRegistration";
     }
     @PostMapping("/removeDetailSale")
-    public String eliminarProducto(@RequestParam("index") int index, Model model) {
+    public String eliminarProducto(@RequestParam("index") Long index, Model model) {
         saleDetailsList.remove(index);
     return "redirect:/finishSale";
 }
